feat: add desktop and mobile footer components
