--- src/components/Footer/FooterMobile.tsx
@@ -1,7 +1,0 @@
-export function FooterMobile() {
-  return (
-    <footer className="h-[115px] flex items-center justify-center border-t border-t-dark-gray">
-      <p className="text-warm-gray">Mobile Footer</p>
-    </footer>
-  )
-}
