--- src/components/Footer/index.ts
@@ -1,2 +1,0 @@
-export * from './FooterDesktop'
-export { FooterMobile } from './FooterMobile'
